VueTesting: Move tests setup to corresponding folder, reuse it across different parts of UTs and ITs
