+++ tests/integration/pages/start-page.spec.ts
@@ -1,49 +1,20 @@
 import Films from '@/vuex/modules/films/state';
-import FiltersPlugin from '@/plugins/filtersPlugin';
-import MockRoutes from '@/mocks/mockEndpoints';
 import StartPage from '@/pages/start-page/StartPage.vue';
-import Vue from 'vue';
-import VueLazyload from 'vue-lazyload';
-import VueResource from 'vue-resource';
 import Vuex from 'vuex';
+import store from '@/vuex/store';
 import { FilmGetterProps } from '@/vuex/modules/films/getters';
 import { FilmsModuleName } from '@/vuex/store/constants';
 import { NoFilmsFoundMessage } from '@/constants/search';
 import { State } from '@/vuex/store/state';
-import { createLocalVue, mount } from '@vue/test-utils';
+import { createDefaultVueInstance } from '../../setup';
 import { delay } from '@/utils';
-import { getAppSettings } from '@/utils/appSettings';
+import { mount } from '@vue/test-utils';
 
-const localVue = createLocalVue();
-localVue.use(VueResource);
-localVue.use(Vuex);
-localVue.use(FiltersPlugin);
-localVue.use(VueLazyload);
-
-localVue.http.options.root = getAppSettings().apiUrl;
-
-// todo: try to find the other way to set "http" to global Vue
-Vue.http = localVue.http;
-
-localVue.http.interceptors.push(((request, next) => {
-    const route = MockRoutes.find(item => {
-        return request.method === item.method && request.url === item.url;
-    });
-
-    if (route) {
-        next(request.respondWith(JSON.stringify(route.response), { status: 200 }));
-    } else {
-        next(request.respondWith({ status: 404, statusText: 'Oh no! Not found!' }));
-    }
-}) as VueResource.HttpInterceptor);
+const localVue = createDefaultVueInstance();
 
 describe('StartPage.vue', () => {
     it('Should render', () => {
         //Arrange
-        const store = new Vuex.Store({
-            modules: { [FilmsModuleName]: Films },
-        });
-
         const wrapper = mount(StartPage, {
             store,
             localVue,
@@ -55,10 +26,6 @@ describe('StartPage.vue', () => {
 
     it('Should display all available films previews in body section from store', () => {
         //Arrange
-        const store = new Vuex.Store({
-            modules: { [FilmsModuleName]: Films },
-        });
-
         const wrapper = mount(StartPage, {
             store,
             localVue,
@@ -106,10 +73,6 @@ describe('StartPage.vue', () => {
     it(`Should display film full description if it was selected from previews`, async () => {
         //Arrange
         const requestDelay: number = 100;
-
-        const store = new Vuex.Store({
-            modules: { [FilmsModuleName]: Films },
-        });
 
         const wrapper = mount(StartPage, {
             store,
